// Iserv.Niis.Angular/src/app/subjects/components/subject-create-dialog/fields.enum.ts
export enum Fields {
  Role = 'roleCode',
  IIN = 'xin',
  NameRu = 'nameRu',
  NameEn = 'nameEn',
  NameKz = 'nameKz',
  CustomerType = 'typeId',
  RegisterNumber = 'registerNumber',
  Status = 'status',
  BeginDate = 'dateBegin',
  EndDate = 'dateEnd',
  Resident = 'resident',
  CountryCode = 'countryCode',
  Country = 'republic',
  Region = 'oblast',
  City = 'city',
  District = 'region',
  Street = 'street',
  AddressRu = 'address',
  AddressEn = 'addressEn',
  AddressKz = 'addressKz',
  Category = 'beneficiaryTypeId',
  OrganizationNameRu = 'organizationNameRu',
  OrganizationNameKz = 'organizationNameKz',
  DirectorName = 'directorName',
  Tel = 'phone',
  MobileTel = 'mobilePhone',
  Fax = 'phoneFax',
  Email = 'email',
  NotNotify = 'isNotMention'
}